fix: ssr/csr 오류 해결 #102
